240521 01:09 jeongsol sweetalert

+++ final_05/src/components/hotplace/HotplaceDetail.vue
@@ -3,6 +3,7 @@ import { ref, onMounted } from "vue";
 import { useRoute, useRouter } from "vue-router";
 import { detailArticle, deleteArticle } from "@/api/hotplace";
 import VArticle from "@/components/common/VArticle.vue";
+import Swal from "sweetalert2";
 const userId = sessionStorage.getItem("userId");
 
 const route = useRoute();
@@ -37,18 +38,41 @@ function moveModify() {
 }
 
 function onDeleteArticle() {
-  deleteArticle(
-    articleno,
-    (response) => {
-      if (response.status == 200) {
-        alert("글이 삭제 되었습니다.");
-        moveList();
-      }
-    },
-    (error) => {
-      console.log(error);
+  Swal.fire({
+    title: "게시글을 삭제하시겠습니까?",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#d33",
+    cancelButtonColor: "#848484",
+    confirmButtonText: "삭제",
+    cancelButtonText: "취소",
+  }).then((result) => {
+    if (result.isConfirmed) {
+      deleteArticle(
+        articleno,
+        (response) => {
+          let msg = response.data;
+          if (response.status == 200) {
+            msg = "게시글이 삭제되었습니다.";
+            Swal.fire({
+              title: "삭제",
+              text: msg,
+              icon: "success",
+            });
+            moveList();
+          } else {
+            Swal.fire({
+              icon: "error",
+              text: msg,
+            });
+          }
+        },
+        (error) => {
+          console.log(error);
+        }
+      );
     }
-  );
+  });
 }
 function getImageUrl(article) {
   if (article && article.fileInfo) {
